update go and rust implementation and push cmd
