+++ .github/workflows/release.yml
@@ -5,54 +5,74 @@ on:
     tags:
       - 'v*'
 
+permissions:
+  contents: write
+
 jobs:
-  release:
-    name: Create Release
+  goreleaser:
+    name: Build Go Binaries
     runs-on: ubuntu-latest
-    permissions:
-      contents: write
-      
     steps:
-    - uses: actions/checkout@v4
-      with:
-        fetch-depth: 0
-        
-    - name: Get version from tag
-      id: get_version
-      run: echo "VERSION=${GITHUB_REF#refs/tags/v}" >> $GITHUB_OUTPUT
-      
-    - name: Extract changelog
-      id: changelog
-      run: |
-        VERSION="${{ steps.get_version.outputs.VERSION }}"
-        CHANGELOG=$(sed -n "/## \[$VERSION\]/,/## \[/p" CHANGELOG.md | sed '1d;$d')
-        echo "CHANGELOG<<EOF" >> $GITHUB_OUTPUT
-        echo "$CHANGELOG" >> $GITHUB_OUTPUT
-        echo "EOF" >> $GITHUB_OUTPUT
-        
-    - name: Create Release
-      uses: actions/create-release@v1
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      with:
-        tag_name: ${{ github.ref }}
-        release_name: Release ${{ steps.get_version.outputs.VERSION }}
-        body: |
-          ## Changes
-          ${{ steps.changelog.outputs.CHANGELOG }}
-          
-          ## Installation
-          
-          ```bash
-          # Clone the repository
-          git clone https://github.com/epcim/git-cross.git
-          cd git-cross
-          
-          # Or vendor it using git-cross itself
-          just cross use git-cross https://github.com/epcim/git-cross.git
-          just cross patch git-cross:. vendor/git-cross
-          ```
-          
-          See [README.md](https://github.com/epcim/git-cross#readme) for full documentation.
-        draft: false
-        prerelease: false
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: '1.23'
+          cache-path: src-go
+      - name: Run GoReleaser
+        uses: goreleaser/goreleaser-action@v6
+        with:
+          distribution: goreleaser
+          version: latest
+          args: release --clean
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+  rust-build:
+    name: Build Rust Binaries
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        include:
+          - os: ubuntu-latest
+            target: x86_64-unknown-linux-gnu
+            artifact_name: git-cross-rust-linux-amd64
+          - os: macos-latest
+            target: x86_64-apple-darwin
+            artifact_name: git-cross-rust-darwin-amd64
+          - os: macos-latest
+            target: aarch64-apple-darwin
+            artifact_name: git-cross-rust-darwin-arm64
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Install Rust
+        uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: ${{ matrix.target }}
+      - name: Build
+        run: cargo build --manifest-path src-rust/Cargo.toml --release --target ${{ matrix.target }}
+      - name: Upload Artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: ${{ matrix.artifact_name }}
+          path: src-rust/target/${{ matrix.target }}/release/git-cross-rust
+
+  upload-rust-to-release:
+    name: Upload Rust to GitHub Release
+    needs: [goreleaser, rust-build]
+    runs-on: ubuntu-latest
+    steps:
+      - name: Download Artifacts
+        uses: actions/download-artifact@v4
+        with:
+          pattern: git-cross-rust-*
+          path: artifacts
+      - name: Upload to Release
+        uses: softprops/action-gh-release@v2
+        if: startsWith(github.ref, 'refs/tags/')
+        with:
+          files: artifacts/**/*
