+++ .gitignore
@@ -10,6 +10,7 @@ testdir
 !.github
 !.specify
 !.gitignore
+!.goreleaser.yaml
 
 
 # Added by cargo
