+++ src-rust/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "git-cross-rust"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = "1.0"
+clap = { version = "4.4", features = ["derive"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tabled = "0.15"
+shell-words = "1.1"
+git2 = { version = "0.18", features = ["vendored-libgit2"] }
+duct = "0.13"
